% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cpr_as_sf}
\alias{cpr_as_sf}
\title{Convert CPR data to sf}
\usage{
cpr_as_sf(
  x,
  coords = c("Longitude (degrees)", "Latitude (degrees)"),
  crs = 4326
)
}
\arguments{
\item{x}{tibble of data}

\item{coords}{char, the names of the x and y coordinates}

\item{crs}{crs to assign during transform}
}
\value{
sf class POINT
}
\description{
Convert CPR data to sf
}
