% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprdata-package.R
\docType{package}
\name{cprdata}
\alias{cprdata}
\alias{cprdata-package}
\title{Continuous Plankton Recorder}
\description{
R language package to assist with access to CPR data.
}
\details{
cprdata package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/cprdata}
}

}
