% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_path}
\alias{get_path}
\title{Get the data path}
\usage{
get_path(filename = "~/.nfsc-cpr")
}
\arguments{
\item{filename}{char, the file where the data path is stored.
Best to allow this to be the default.}
}
\value{
the data path
}
\description{
Get the data path
}
