% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{harmonize_cpr}
\alias{harmonize_cpr}
\title{Harmonize GMRI and NFSC CPR datasets}
\usage{
harmonize_cpr(
  x = read_nfsc_cpr(),
  y = read_gmri_cpr(),
  is_zoo = "taxon_stage" \%in\% names(y)
)
}
\arguments{
\item{x}{long-form table of NFSC data}

\item{y}{table of GMRI data}

\item{is_zoo}{logical, if TRUE then account for taxon stage}
}
\value{
list of two tables with the same form, one for nfsc and gmri
}
\description{
Harmonize GMRI and NFSC CPR datasets
}
