% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_cpr}
\alias{import_cpr}
\title{Import data from a NFSC-cpr spreadsheet}
\usage{
import_cpr(filename = list_files(full.names = TRUE)[1])
}
\arguments{
\item{filename}{char, the name of the file to read}
}
\value{
list of tibbles with one sheet per element
}
\description{
Import data from a NFSC-cpr spreadsheet
}
