% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_files}
\alias{list_files}
\title{List the data files stored in the path}
\usage{
list_files(pattern = "^.*\\\\.xlsx$", path = get_path(), ...)
}
\arguments{
\item{pattern}{char, the regex pattern to search for.  Use
\code{glob2rx()} to convert globs to regular expressions.}

\item{path}{char the data path}

\item{...}{other arguments for \code{list.files()}}
}
\value{
zero or more filenames
}
\description{
List the data files stored in the path
}
