% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{make_stage_lut}
\alias{make_stage_lut}
\title{Make the beginnings of a LUT for managing stage names}
\usage{
make_stage_lut(x = read_cpr("zooplankton", composite = FALSE, clean = FALSE))
}
\arguments{
\item{x}{table of unadultered data}
}
\value{
a table of name, n (count) and newname
}
\description{
This provides and intermediary step, the output "newname" column
should be adjusted, then the table saved as inst/stage/stage_lut.csv
}
