% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{nfsc_cpr_extract_stage}
\alias{nfsc_cpr_extract_stage}
\title{Given a nfsc cpr table split name and stage}
\usage{
nfsc_cpr_extract_stage(x)
}
\arguments{
\item{x}{table, the nfsc cpr data (long format)}
}
\value{
the tabkle with updated name and stage columns
}
\description{
Given a nfsc cpr table split name and stage
}
