% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsccpr-package.R
\docType{package}
\name{nfsccpr}
\alias{nfsccpr}
\alias{nfsccpr-package}
\title{nfsccpr package}
\description{
R language package to assist with access to NFSC CPR data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/nefsc-cpr}
}

}
