% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gmri_cpr}
\alias{read_gmri_cpr}
\title{Read GMRI data downloaded by NERACOOS}
\usage{
read_gmri_cpr(
  filename =
    system.file("gmri/2024-03-08_gom_cpr_zooplankton_full_bdeb_74de_8e13.csv.gz", package
    = "nfsccpr"),
  repair_longitude = TRUE
)
}
\arguments{
\item{filename, }{the name of the file to read}

\item{repair_longitude}{logi, if TRUE then make all longitudes negative}
}
\value{
table of data (with units as an attribute)
}
\description{
Read GMRI data downloaded by NERACOOS
}
