% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_nfsc_cpr}
\alias{read_nfsc_cpr}
\title{Read NFSC data exported freom NFSC Excel spreadsheet}
\usage{
read_nfsc_cpr(
  filename = system.file("nfsc/2024-01-20_zooplankton.csv.gz", package = "nfsccpr")
)
}
\arguments{
\item{filename, }{the name of the file to read}
}
\value{
table of data
}
\description{
Read NFSC data exported freom NFSC Excel spreadsheet
}
