% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{read_stage_lut}
\alias{read_stage_lut}
\title{Read the stage recoding look up table (lut)}
\usage{
read_stage_lut(
  filename = system.file("stage/stage_lut.csv", package = "nfsccpr"),
  form = c("table", "vector")[1]
)
}
\arguments{
\item{filename}{chr , the name of the file to read}

\item{form}{chr, one of "table" or "vector"}
}
\value{
table of name and newname or a charcater named vector of name = newname
}
\description{
Read the stage recoding look up table (lut)
}
