% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_path}
\alias{set_path}
\title{Set the data path}
\usage{
set_path(path = ".", filename = "~/.nfsc-cpr")
}
\arguments{
\item{path}{char, the data path where your XLSX file is stored}

\item{filename}{char, the name of the file where you will store the path.
Best to allow this to be the default.}
}
\value{
the data path
}
\description{
Set the data path
}
