% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{stage_clean}
\alias{stage_clean}
\title{Clean a vector of stages}
\usage{
stage_clean(x)
}
\arguments{
\item{x}{chr a character vector of stages}
}
\value{
the input in sentence-case and spaces minimized
}
\description{
Clean a vector of stages
}
