% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{nfsc_cpr_to_long}
\alias{nfsc_cpr_to_long}
\title{Pivot CPR data to long form}
\usage{
nfsc_cpr_to_long(x, retain_col = "Phytoplankton Color Index")
}
\arguments{
\item{x}{tibble of data}
}
\value{
long-form tibble of data
}
\description{
Pivot CPR data to long form
}
